feat(endpoint-micropub): remove frontend interface
