--- packages/endpoint-micropub/locales/de.json
@@ -1,10 +1,0 @@
-{
-  "micropub": {
-    "posts": {
-      "none": "Keine Beiträge",
-      "title": "Veröffentlichte Beiträge"
-    },
-    "properties": "Eigenschaften",
-    "title": "Beiträge"
-  }
-}
