--- packages/endpoint-micropub/locales/en.json
@@ -1,10 +1,0 @@
-{
-  "micropub": {
-    "title": "Posts",
-    "properties": "Properties",
-    "posts": {
-      "title": "Published posts",
-      "none": "No posts"
-    }
-  }
-}
