--- packages/endpoint-micropub/locales/es.json
@@ -1,10 +1,0 @@
-{
-  "micropub": {
-    "posts": {
-      "none": "No hay publicaciones",
-      "title": "Publicaciones publicadas"
-    },
-    "properties": "Propiedades",
-    "title": "Publicaciones"
-  }
-}
