--- packages/endpoint-micropub/locales/fr.json
@@ -1,10 +1,0 @@
-{
-  "micropub": {
-    "posts": {
-      "none": "Pas d'articles",
-      "title": "Articles publiés"
-    },
-    "properties": "Propriétés",
-    "title": "Articles"
-  }
-}
