--- packages/endpoint-micropub/locales/id.json
@@ -1,10 +1,0 @@
-{
-  "micropub": {
-    "posts": {
-      "none": "Tidak ada postingan",
-      "title": "Postingan yang diterbitkan"
-    },
-    "properties": "Properti",
-    "title": "Postingan"
-  }
-}
