--- packages/endpoint-micropub/locales/nl.json
@@ -1,10 +1,0 @@
-{
-  "micropub": {
-    "posts": {
-      "none": "Geen posts",
-      "title": "Gepubliceerde berichten"
-    },
-    "properties": "Eigenschappen",
-    "title": "Berichten"
-  }
-}
