--- packages/endpoint-micropub/locales/pt.json
@@ -1,10 +1,0 @@
-{
-  "micropub": {
-    "posts": {
-      "none": "Nenhuma postagem",
-      "title": "Postagens publicadas"
-    },
-    "properties": "Propriedades",
-    "title": "Postagens"
-  }
-}
