+++ packages/endpoint-micropub/package.json
@@ -43,7 +43,6 @@
     "lodash": "^4.17.21",
     "markdown-it": "^13.0.0",
     "microformats-parser": "^1.4.0",
-    "mongodb": "^4.1.1",
     "multer": "^1.4.5-lts",
     "newbase60": "^1.3.1",
     "turndown": "^7.1.1",
