--- packages/endpoint-micropub/tests/integration/501-posts-list-no-database.js
@@ -1,23 +1,0 @@
-import process from "node:process";
-import test from "ava";
-import { JSDOM } from "jsdom";
-import { testServer } from "@indiekit-test/server";
-
-test("Returns list of previously uploaded files", async (t) => {
-  const request = await testServer({ useDatabase: false });
-  const response = await request
-    .get("/micropub")
-    .auth(process.env.TEST_TOKEN, { type: "bearer" });
-  const dom = new JSDOM(response.text);
-
-  const result = dom.window.document;
-
-  t.is(
-    result.querySelector("title").textContent,
-    "Not Implemented - Test configuration"
-  );
-  t.is(
-    result.querySelector(".article__body p").textContent,
-    "This feature requires a database."
-  );
-});
