--- packages/endpoint-micropub/views/post.njk
@@ -1,24 +1,0 @@
-{% extends "document.njk" %}
-
-{% set title = post.properties.name or post.properties["post-type"] | capitalize %}
-{% set content = post.properties.content.html %}
-
-{% block content %}
-  {% if post.properties.audio %}
-    {% for item in post.properties.audio %}
-      <audio src="{{ item.url | absoluteUrl(publication.me) }}" controls>
-    {% endfor %}
-  {% elif post.properties.photo %}
-    {% for item in post.properties.photo %}
-      <img src="{{ item.url | absoluteUrl(publication.me) }}" alt="{{ item.alt }}">
-    {% endfor %}
-  {% elif post.properties.video %}
-    {% for item in post.properties.video %}
-      <video src="{{ item.url | absoluteUrl(publication.me) }}" controls>
-    {% endfor %}
-  {% endif %}
-  {{ summary({
-    title: __("micropub.properties"),
-    rows: post.properties | summaryRows
-  }) }}
-{% endblock %}
