--- packages/endpoint-micropub/views/posts.njk
@@ -1,12 +1,0 @@
-{% extends "document.njk" %}
-
-{% block content %}
-  {%- if posts.length > 0 %}
-  {{ fileList({
-    items: posts
-  }) }}
-  {{ pagination(pages(page, limit, count)) }}
-  {%- else -%}
-  <p>{{ __("micropub.posts.none") }}<p>
-  {%- endif %}
-{% endblock %}
